backend/app/requirements.txt
fastapi
uvicorn[standard]
boto3
pillow
imagehash
python-multipart
pydantic
requests
